Moved registry outside of ml_worker module
