+++ docs/reference/transformation-functions/index.rst
@@ -5,7 +5,7 @@ Transformation functions
 
 .. autofunction:: giskard.transformation_function
 
-.. autoclass:: giskard.ml_worker.testing.registry.transformation_function.TransformationFunction
+.. autoclass:: giskard.registry.transformation_function.TransformationFunction
 
    .. automethod:: execute
    .. automethod:: upload
