+++ giskard/datasets/base/__init__.py
@@ -23,8 +23,8 @@ from giskard.client.python_utils import warning
 from giskard.core.core import NOT_GIVEN, DatasetMeta, NotGivenOr, SupportedColumnTypes
 from giskard.core.errors import GiskardImportError
 from giskard.core.validation import configured_validate_arguments
-from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction, SlicingFunctionType
-from giskard.ml_worker.testing.registry.transformation_function import (
+from giskard.registry.slicing_function import SlicingFunction, SlicingFunctionType
+from giskard.registry.transformation_function import (
     TransformationFunction,
     TransformationFunctionType,
 )
