+++ giskard/ml_worker/ml_worker.py
@@ -12,7 +12,7 @@ from giskard.core.validation import ConfiguredBaseModel
 from giskard.ml_worker.stomp.client import StompWSClient
 from giskard.ml_worker.stomp.constants import HeaderType
 from giskard.ml_worker.stomp.parsing import Frame, StompFrame
-from giskard.ml_worker.testing.registry.registry import load_plugins
+from giskard.registry.registry import load_plugins
 from giskard.ml_worker.websocket.action import ActionPayload, ConfigPayload, MLWorkerAction
 from giskard.ml_worker.websocket.listener import WEBSOCKET_ACTORS, MLWorkerInfo
 from giskard.ml_worker.websocket.utils import fragment_message
