+++ giskard/ml_worker/testing/functions/transformation.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 from scipy.stats import median_abs_deviation
 
-from giskard.ml_worker.testing.registry.transformation_function import transformation_function
+from giskard.registry.transformation_function import transformation_function
 
 from ....llm import get_default_client
