+++ giskard/push/push_test_catalog/catalog.py
@@ -17,8 +17,8 @@ from typing import Any, Optional
 
 from giskard.datasets.base import Dataset
 from giskard.ml_worker.testing.functions.transformation import mad_transformation
-from giskard.ml_worker.testing.registry.decorators import test
-from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction
+from giskard.registry.decorators import test
+from giskard.registry.slicing_function import SlicingFunction
 from giskard.ml_worker.testing.test_result import TestResult
 from giskard.models.base import BaseModel
 from giskard.testing.tests.calibration import test_overconfidence_rate, test_underconfidence_rate
