+++ giskard/registry/decorators.py
@@ -4,8 +4,8 @@ import sys
 from typing import Callable, Optional, List, Union, Type, TypeVar
 
 from giskard.core.core import TestFunctionMeta
-from giskard.ml_worker.testing.registry.decorators_utils import make_all_optional_or_suite_input, set_return_type
-from giskard.ml_worker.testing.registry.giskard_test import GiskardTestMethod, GiskardTest
+from giskard.registry.decorators_utils import make_all_optional_or_suite_input, set_return_type
+from giskard.registry.giskard_test import GiskardTestMethod, GiskardTest
 
 
 def test(
@@ -28,7 +28,7 @@ def test(
         Declare output as both Callable and GiskardTest so that there's autocompletion
         for GiskardTest's methods as well as the original wrapped function arguments (for __call__)
         """
-        from giskard.ml_worker.testing.registry.registry import tests_registry
+        from giskard.registry.registry import tests_registry
 
         tests_registry.register(
             TestFunctionMeta(original, name=name, tags=tags, debug_description=debug_description, type="TEST")
