+++ giskard/registry/transformation_function.py
@@ -7,13 +7,13 @@ from typing import Callable, List, Optional, Set, Type, Union
 from giskard.core.core import DatasetProcessFunctionMeta
 from giskard.core.savable import Artifact, RegistryArtifact
 from giskard.core.validation import configured_validate_arguments
-from giskard.ml_worker.testing.registry.decorators_utils import (
+from giskard.registry.decorators_utils import (
     drop_arg,
     make_all_optional_or_suite_input,
     set_return_type,
     validate_arg_type,
 )
-from giskard.ml_worker.testing.registry.registry import get_object_uuid, tests_registry
+from giskard.registry.registry import get_object_uuid, tests_registry
 
 TransformationFunctionType = Callable[..., Union[pd.Series, pd.DataFrame]]
 
@@ -115,7 +115,7 @@ def transformation_function(
     """
 
     def inner(func: Union[TransformationFunctionType, Type[TransformationFunction]]) -> TransformationFunction:
-        from giskard.ml_worker.testing.registry.registry import tests_registry
+        from giskard.registry.registry import tests_registry
 
         tests_registry.register(
             DatasetProcessFunctionMeta(
