+++ giskard/testing/tests/calibration.py
@@ -4,8 +4,8 @@ import numpy as np
 import pandas as pd
 
 from ...datasets.base import Dataset
-from ...ml_worker.testing.registry.decorators import test
-from ...ml_worker.testing.registry.slicing_function import SlicingFunction
+from ...registry.decorators import test
+from ...registry.slicing_function import SlicingFunction
 from ...ml_worker.testing.test_result import TestResult
 from ...models.base import BaseModel
 from . import debug_description_prefix
