+++ giskard/testing/tests/debug_slicing_functions.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from giskard.ml_worker.testing.registry.slicing_function import slicing_function
+from giskard.registry.slicing_function import slicing_function
 
 
 # Performance tests: Classification
