+++ giskard/testing/tests/drift.py
@@ -12,8 +12,8 @@ from scipy.stats.stats import Ks_2sampResult, wasserstein_distance
 from typing import List, Optional
 
 from giskard.datasets.base import Dataset
-from giskard.ml_worker.testing.registry.decorators import test
-from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction
+from giskard.registry.decorators import test
+from giskard.registry.slicing_function import SlicingFunction
 from giskard.ml_worker.testing.test_result import TestMessage, TestMessageLevel, TestResult
 from giskard.models.base import BaseModel
 from giskard.testing.utils.utils import check_slice_not_empty, validate_classification_label
