+++ giskard/testing/tests/llm/hallucination.py
@@ -3,7 +3,7 @@ from typing import Optional
 from ....datasets.base import Dataset
 from ....llm.evaluators.coherency import CoherencyEvaluator
 from ....llm.evaluators.plausibility import PlausibilityEvaluator
-from ....ml_worker.testing.registry.decorators import test
+from ....registry.decorators import test
 from ....ml_worker.testing.test_result import TestResult
 from ....models.base.model import BaseModel
