+++ giskard/testing/tests/llm/injections.py
@@ -11,7 +11,7 @@ from ....datasets.base import Dataset
 from ....llm import LLMImportError
 from ....llm.prompt_injection.data import Prompt, StringMatchingMethod
 from ....llm.prompt_injection.evaluator import evaluate as evaluate_pi
-from ....ml_worker.testing.registry.decorators import test
+from ....registry.decorators import test
 from ....ml_worker.testing.test_result import TestResult
 from ....models.base import BaseModel
 from .. import debug_description_prefix
