+++ giskard/testing/tests/statistic.py
@@ -5,8 +5,8 @@ import numpy as np
 from typing import Optional
 
 from giskard.datasets.base import Dataset
-from giskard.ml_worker.testing.registry.decorators import test
-from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction
+from giskard.registry.decorators import test
+from giskard.registry.slicing_function import SlicingFunction
 from giskard.ml_worker.testing.test_result import TestMessage, TestMessageLevel, TestResult
 from giskard.models.base import BaseModel
 from giskard.testing.utils.utils import check_slice_not_empty, validate_classification_label
