+++ tests/debug/test_statistic.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 
-from giskard.ml_worker.testing.registry.slicing_function import slicing_function
+from giskard.registry.slicing_function import slicing_function
 from giskard.testing.tests.statistic import test_right_label, test_output_in_range, test_disparate_impact
