+++ tests/test_data_processing_pipeline.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pytest
 
 from giskard import slicing_function, Dataset, SlicingFunction
-from giskard.ml_worker.testing.registry.transformation_function import transformation_function
+from giskard.registry.transformation_function import transformation_function
 
 
 @slicing_function(name="slice with parenthesis")
