+++ tests/testing/test_test_metadata.py
@@ -5,8 +5,8 @@ import pandas as pd
 import pytest
 
 from giskard import test, slicing_function, transformation_function
-from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction
-from giskard.ml_worker.testing.registry.transformation_function import TransformationFunction
+from giskard.registry.slicing_function import SlicingFunction
+from giskard.registry.transformation_function import TransformationFunction
 from giskard.ml_worker.testing.test_result import TestResult as GiskardTestResult
 from giskard.models.base import BaseModel
